+++ B3.Domain/Interfaces/IDescontoImpostoRendaService.cs
@@ -5,7 +5,7 @@ namespace B3.Domain.Interfaces
     public interface IDescontoImpostoRendaService
     {
         Task<List<DescontoImpostoRenda>> GetDescontoImpostoRendas();
-        Task<float> CalcularDescontoIR(float valorRendimento, int qtdeMesesInvestimento);
+        Task<decimal> CalcularDescontoIR(decimal valorRendimento, int qtdeMesesInvestimento);
 
     }
 }
