+++ B3.Domain/ViewModels/SimulacaoTituloViewModel.cs
@@ -2,10 +2,10 @@
 {
     public class SimulacaoTituloViewModel
     {
-        public float ValorTotalBruto { get; set; }
-        public float ValorTotalInvestido { get; set; }
-        public float ValorRendimento { get; set; }
-        public float ValorDescontoImpostoRenda { get; set; }
-        public float ValorTotalLiquido { get; set; }
+        public decimal ValorTotalBruto { get; set; }
+        public decimal ValorTotalInvestido { get; set; }
+        public decimal ValorRendimento { get; set; }
+        public decimal ValorDescontoImpostoRenda { get; set; }
+        public decimal ValorTotalLiquido { get; set; }
     }
 }
